+++ data/contactForm.ts
@@ -22,7 +22,8 @@ export const CONTACT_VALIDATION_MESSAGE_REQUIRED = "Add a short message";
 
 export const CONTACT_FORM_SUBMIT_IDLE = "Send message";
 export const CONTACT_FORM_SUBMIT_LOADING = "Sending…";
-export const CONTACT_FORM_SUCCESS_MESSAGE = "Message sent. I will reply within a few business days.";
+export const CONTACT_FORM_SUCCESS_MESSAGE =
+  "Message sent. I will reply within a few business days.";
 
 export const CONTACT_FORM_ERROR_GENERIC = "Something went wrong. Please try again.";
 export const CONTACT_FORM_ERROR_NETWORK = "Failed to send. Please try again or email directly.";